+++ src/pages/RegisterEachMedi/components/RegisterEachMediForm.tsx
@@ -1,5 +1,5 @@
 import React, { useState, forwardRef } from "react";
-import instanceWithToken from "../../../apis/axiosInstance";
+import instanceWithToken from "../../../apis/axiosInstanceWithToken";
 import { ReactComponent as Down } from "../../../assets/registerMedi/down.svg";
 import AutoMediInput from "../../../components/Common/AutoMediInput";
